Refactored Url check
I have removed the writing to file for URL Check to meet the S in the SOLID principles.
