+++ internal/platform/directory/directory.go
@@ -2,6 +2,8 @@ package directory
 
 import (
 	"fmt"
+	"io"
+	"log"
 	"os"
 )
 
@@ -19,4 +21,17 @@ func CreateDirectory(path string) error {
 func GetFilePathTemplate(owner string, repoName string) string {
 	filePathTemplate := fmt.Sprintf("./docs/%s/%s/", owner, repoName)
 	return filePathTemplate
+}
+
+func OutputToFile(path string, link string, code string) error {
+	f, err := os.OpenFile(path+"output.txt", os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0600)
+	if err != nil {
+		log.Printf("Failed to create output file: %v\n", err)
+	}
+
+	defer f.Close()
+
+	_, err = io.WriteString(f, link+" - "+code+"\n")
+
+	return nil
 }
